refactor(#653): consolidate structural invariant tests into architecturalInvariants.test.ts
Consolidate 8 structural source-scanning tests from 4 files into a single
architecturalInvariants.test.ts with shared helpers for store file scanning,
brace-counted function body extraction, and syncQueue pattern detection.

Each invariant is documented with the SEV/incident it guards against:
- Invariant 1: Delete routing discipline (SEV1 2026-04-12)
- Invariant 2: SyncQueue idempotency (no .insert() in retry path)
- Invariant 3: _fetchFromSupabase READ path is read-only (SEV1 2026-04-12)
- Invariant 4: Cross-tab sync completeness

Extracted shared helpers: getStoreFilePaths(), getStoreFiles(),
extractFunctionBody(), findPatternInsideSyncEnqueue().

Original files cleaned up: syncQueueSafety.test.ts and
crossTabSyncStructural.test.ts now contain redirect stubs;
syncQueue.test.ts and workout.test.ts retain their behavioral tests.
