+++ src/lib/__tests__/syncQueue.test.ts
@@ -446,68 +446,5 @@ describe('SyncQueue — delete circuit breaker', () => {
   })
 })
 
-// ─────────────────────────────────────────────────────────────────
-// Structural guard: every Supabase DELETE operation routed through
-// syncQueue must go through enqueueDelete (not plain enqueue), so
-// the circuit breaker sees it.
-// ─────────────────────────────────────────────────────────────────
-
-describe('SyncQueue — delete routing discipline (structural)', () => {
-  it('stores never wrap a .delete() in plain syncQueue.enqueue', async () => {
-    const { readFileSync } = await import('node:fs')
-    const { resolve } = await import('node:path')
-    const { fileURLToPath } = await import('node:url')
-    const { dirname } = await import('node:path')
-
-    const __filename = fileURLToPath(import.meta.url)
-    const __dirname = dirname(__filename)
-    const storeDir = resolve(__dirname, '../../stores')
-
-    const files = ['workout.ts', 'bodyweight.ts', 'preferences.ts', 'progression.ts']
-    for (const file of files) {
-      const src = readFileSync(resolve(storeDir, file), 'utf-8')
-      // Find every syncQueue.enqueue( and check a window of the following 300 chars
-      // for .delete() — if present, it should have been enqueueDelete instead.
-      const enqueueRe = /syncQueue\.enqueue\(/g
-      let match: RegExpExecArray | null
-      while ((match = enqueueRe.exec(src)) !== null) {
-        const window = src.slice(match.index, match.index + 300)
-        if (/\.delete\s*\(/.test(window)) {
-          throw new Error(
-            `${file} has syncQueue.enqueue wrapping a .delete() at offset ${match.index}. Use syncQueue.enqueueDelete so the circuit breaker sees it. Context:\n${window.slice(0, 200)}`,
-          )
-        }
-      }
-    }
-  })
-
-  // Gate 5 invariant: stores do not issue hard DELETEs at all.
-  // Every removal must go through UPDATE { deleted_at: ... } so data is
-  // recoverable within the grace window before the hard-delete cron runs.
-  it('stores never call .delete() on a Supabase query — all removals are soft', async () => {
-    const { readFileSync } = await import('node:fs')
-    const { resolve } = await import('node:path')
-    const { fileURLToPath } = await import('node:url')
-    const { dirname } = await import('node:path')
-
-    const __filename = fileURLToPath(import.meta.url)
-    const __dirname = dirname(__filename)
-    const storeDir = resolve(__dirname, '../../stores')
-
-    const files = ['workout.ts', 'bodyweight.ts', 'preferences.ts', 'progression.ts']
-    for (const file of files) {
-      const src = readFileSync(resolve(storeDir, file), 'utf-8')
-      // Match any supabase query builder .delete(), allowing whitespace / newlines
-      // between .from(...) and .delete(). Non-supabase .delete() (Map/Set) is not
-      // matched because the anchor requires .from(...) upstream within 200 chars.
-      const re = /\.from\(\s*['"][^'"]+['"]\s*\)[\s\S]{0,200}?\.delete\s*\(/g
-      const match = re.exec(src)
-      if (match) {
-        const offset = match.index
-        throw new Error(
-          `${file} contains a hard .delete() on a Supabase query at offset ${offset}. Gate 5 requires UPDATE { deleted_at: new Date().toISOString() } for all removals. Context:\n${src.slice(offset, offset + 200)}`,
-        )
-      }
-    }
-  })
-})
+// Structural invariant tests (delete routing discipline, Gate 5 soft-delete)
+// have been consolidated into architecturalInvariants.test.ts (LIFT-653).
